requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
psycopg2
pandas
celery
rabbitmq
redis
pytest
pytest_django
requests
drf_yasg